stm32f407dmaAdc
Everything in ADC configurations will be as default in normal mode. However, this time the ADC interrupts are not activated and the DMA is configured instead and DMA interrupt is enabled by default in the NVIC controller tab. The DMA configurations for the ADC will be as shown down below. Just add a DMA channel and that’s all.